% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_score.R
\name{adjacency_score}
\alias{adjacency_score}
\title{Ranks pairs of features according to the Adjacency Score}
\usage{
adjacency_score(g2, f, f_pairs, num_perms = 1000, num_cores = 1)
}
\arguments{
\item{g2}{an object of the class \code{simplicial} containing the nerve or clique complex.}

\item{f}{a numeric vector or matrix specifying one or more functions with support on
the set of points whose significance will be assesed in the simplicial complex. Each
column corresponds to a point and each row specifies a different function.}

\item{f_pairs}{a 2 column matrix where each row specifies the row names
of two functions or features to be assessed}

\item{num_perms}{number of permutations used to build the null distribution for each
feature. By default is set to 1000.}

\item{num_cores}{integer specifying the number of cores to be used in the computation. By
default only one core is used.}
}
\value{
Returns a data frame with the value of the Adjacency Score,
the p-values, and the q-values computed using Benjamini-Hochberg procedure.
}
\description{
Uses the Laplacian of a nerve or clique complex to assess the significant colocalization of
given pairs of features defined over the points in that complex. Significance is computed by
comparing to a null distribution created by reshuffling the features across the points. Pairs
of features are permuted in the same way.
}
