% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmarks feature selection algorithms using various datasets.}
\usage{
benchmark(sorted_results, q, r = 10, num_cores = 1, dataset = "gisette")
}
\arguments{
\item{sorted_results}{a dataframe with each row corresponding to a different feature. Rows are
ranked according to decreasing significance.}

\item{q}{an integer or a vector of integers specifying the number of features taken into account
in the evaluation. If a list of integers is specified, benchmarks are sequentially performed for
the different values in the list.}

\item{r}{an integer specifying the number of times the benchmark is repeated for each value of
\code{q}. By default is set to 10.}

\item{num_cores}{number of cores used for the computation. By default is set to 1.}

\item{dataset}{a string specifying the dataset used for the benchmark. Supported datasets are
\code{gisette} (default), \code{dexter}, and \code{mnist}.}
}
\value{
Returns a matrix with \code{r} rows and \code{q} columns containing the normalized mutual
information estimated from the empirical probability distribution. In addition, the average mutual
information is ploted as a function of \code{q}.
}
\description{
Benchmarks a ranked list of features using a k-means classifier based on the top \code{q} features
in the list. The normalized mutual information is used as a metric.
}
\examples{
library(RayleighSelection)
# Load GISETTE dataset and an example of ranked list produced using \\code{rayleigh_selection}
data("gisette")
data("gisette_results")

# Rank pixels according to their p-value
sorted_results <- gisette_results[order(gisette_results$p),]

# Benchmarks the ranked list of pixels
benchmark(sorted_results, seq(500, length(row.names(sorted_results)), 1000))

}
