% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_laplacian.R
\name{combinatorial_laplacian}
\alias{combinatorial_laplacian}
\title{Computes the Combinatorial Laplacian for 0- and 1-forms.}
\usage{
combinatorial_laplacian(g2, one_forms = TRUE, weights = FALSE)
}
\arguments{
\item{g2}{an object of the class \code{simplicial} containing the nerve or clique complex.}

\item{one_forms}{when set FALSE only the Combinatorial Laplacian for 0-forms is
computed. By default is set to TRUE.}

\item{weights}{when set to TRUE it takes 2-simplices into account when computing weights.
By default is set to FALSE.}
}
\value{
Returns a list with the Combinatorial Laplacians for 0- and 1-forms
and the weights for 0- and 1-simplices.
}
\description{
Given a nerve or a clique complex, it computes the Combinatorial Laplacians for 0- and 1-forms and 
the weights of 0- and 1-simplices.
}
\examples{
library(RayleighSelection)
gy <- nerve_complex(list(c(1,4,6,10), c(1,2,7), c(2,3,8), c(3,4,9,10), c(4,5)))
combinatorial_laplacian(gy)

}
