% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_to_complex.R
\name{graph_to_complex}
\alias{graph_to_complex}
\title{Generates a simplicial complex from a graph.}
\usage{
graph_to_complex(adjacency, clique = TRUE)
}
\arguments{
\item{adjacency}{a weighted adjacency matrix.}

\item{clique}{if set to FALSE the computation of 2-simplices is skipped.}
}
\value{
An object of the class \code{simplicial}. The class \code{simplicial} inherits from
the class \code{igraph}.
}
\description{
Takes the adjacency matrix of a graph and returns a clique complex that can be used
by other functions in the \code{RayleighSelection} package.
}
\examples{
library(RayleighSelection)
# Load MNIST example dataset
data("mnist")

# Compute a correlation matrix for a subset of the LFW dataset using only pixels with high variance
mnist_test <- mnist[,1:800]
mnist_test_top <- mnist_test[apply(mnist_test, 1, var) > 0.9,]
mnist_test_distances <- cor(mnist_test_top)

# Compute a simplifical complex
gg <- graph_to_complex(mnist_test_distances)

# Plot the skeleton of the simplicial complex colored by the intensity of the 500th pixel
plot_skeleton(gg, k=as.numeric(mnist_test[500,]))

}
