% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_complex.R
\name{nerve_complex}
\alias{nerve_complex}
\title{Generates the nerve complex of an open cover.}
\usage{
nerve_complex(open_cover)
}
\arguments{
\item{open_cover}{a list of open sets with \code{open_cover[[i]]} a vector of indices of points
corresponding to the \code{i}-th open set in the open cover.}
}
\value{
An object of the class \code{simplicial}. The class \code{simplicial} inherits from
the class \code{igraph}.
}
\description{
Takes the open cover of a finite set and returns its nerve
complex. The output is a simplicial complex that can be used by other functions in the
\code{RayleighSelection} package.
}
