% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_complex.R
\name{knn_complex}
\alias{knn_complex}
\title{Generates a k-nearest neighbor complex from a distance matrix.}
\usage{
knn_complex(dist, k)
}
\arguments{
\item{dist}{a distance matrix.}

\item{k}{nunber of nearest neighbors}
}
\value{
An object of the class \code{simplicial}. The class \code{simplicial} inherits from
the class \code{igraph}.
}
\description{
Takes a distance matrix, computes a k-nearest neighbor graph, and returns
a clique complex that can be used by other functions in the
\code{RayleighSelection} package.
}
\examples{
library(RayleighSelection)
# Load MNIST example dataset
data("mnist")

# Compute a distance matrix for a subset of the MNIST dataset
mnist_test <- mnist[,1:800]
mnist_test_distances <- (1.0 - cor(mnist_test))

# Compute nearest neighbor complex with k = 5
gg <- knn_complex(mnist_test, 5)

# Plot the skeleton of the nearest neighbor graph colored by the intensity of the 500th pixel
plot_skeleton(gg, k=as.numeric(mnist_test[500,]))

}
