% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_complex.R
\name{nerve_complex}
\alias{nerve_complex}
\title{Generates the nerve complex of an open cover.}
\usage{
nerve_complex(open_cover)
}
\arguments{
\item{open_cover}{a list of open sets with \code{open_cover[[i]]} a vector of indices of points
corresponding to the \code{i}-th open set in the open cover.}
}
\value{
An object of the class \code{simplicial}. The class \code{simplicial} inherits from
the class \code{igraph}.
}
\description{
Takes the open cover of a finite set and returns its nerve
complex. The output is a simplicial complex that can be used by other functions in the
\code{RayleighSelection} package.
}
\examples{
library(RayleighSelection)
# Load processed LFW dataset
data("lfw_distances")
data("lfw")

# Compute reduced representation using Laplacian eigenmap
library(dimRed)
leim <- LaplacianEigenmaps()
storage.mode(lfw) <- "double"
emb <- leim@fun(as(t(lfw), "dimRedData"), leim@stdpars)

# Compute Mapper representation using the Laplacian eigenmap as an auxiliary function
library(TDAmapper)
m2 <- mapper2D(distance_matrix = lfw_distances,
               filter_values = list(emb@data@data[,1], emb@data@data[,2]),
               num_intervals = c(40,40),
               percent_overlap = 30,
               num_bins_when_clustering = 10);

# Compute the nerve complex
gg <- nerve_complex(m2$points_in_vertex)

# Plot the skeleton of the nerve complex colored by the intensity of the 1063rd pixel
plot_skeleton(gg, k=as.numeric(lfw[1063,]), spring.constant = 5.0)

}
